% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene.R
\name{multi_gene}
\alias{multi_gene}
\title{multi_gene}
\usage{
multi_gene(
  annotation,
  read,
  readmap,
  novel_transcript_ids = NULL,
  chr,
  range,
  strand = "both",
  limit = 30,
  read_low_color = "#ADD8E6",
  read_high_color = "#4682B4",
  annotation_color = "#990099",
  novel_isoform_color = "#CC0000",
  known_isoform_color = "#00DD00",
  text_color = "#000000",
  text_size = 3,
  text_alpha = 1,
  text_fontface = "bold",
  title_color = "#000000",
  title_size = 10,
  title_face = "bold",
  x_text_size = 2,
  x_text_just = 0.25,
  num_color = "#000000",
  num_size = 3,
  num_alpha = 1,
  num_fontface = "bold",
  num_margin = 100,
  show_transcript_name = FALSE
)
}
\arguments{
\item{annotation}{A gtf format data,such as Homo_sapiens.GRCh38.109.chr.gtf.}

\item{read}{A bed format data after alignment with the genome,having the GRanges object and the metadata of name column.}

\item{readmap}{A tsv format data,mapping the transcript_id to read name;The column name needs to contain the name and transcript id.}

\item{novel_transcript_ids}{A dataframe with the column name needs to contain novel_transcript_id to mark the color of novel isoform.}

\item{chr}{The chromosome of interest,such as chr="6".}

\item{range}{The range of interest,such as range=c(36000000,36700000).}

\item{strand}{The strand of interest,such as strand="+" or "-" or "both";the default value is "both".}

\item{limit}{To limit the number of read displays. If the read count exceeds the limit value, the limit+log2(read count-limit)  number of reads will be randomly selected.}

\item{read_low_color}{The color is the minimum number of reads.}

\item{read_high_color}{The color is the maximum number of reads.}

\item{annotation_color}{The annotation exon color in the gtf format data.}

\item{novel_isoform_color}{Mark the color of the novel transcript isoform.}

\item{known_isoform_color}{Mark the color of the known transcript isoform.}

\item{text_color}{The font color of transcript name.}

\item{text_size}{The font size of transcript name.}

\item{text_alpha}{The font transparency of transcript name.}

\item{text_fontface}{The  font shape of transcript name.}

\item{title_color}{The font color  of tittle name.}

\item{title_size}{The font size of tittle name.}

\item{title_face}{The  font shape of tittle name.}

\item{x_text_size}{The text size of x axis.}

\item{x_text_just}{Adjust the x - axis text position.}

\item{num_color}{The font color of read number.}

\item{num_size}{The font size of read number.}

\item{num_alpha}{The font transparency of read number.}

\item{num_fontface}{The  font shape of read number.}

\item{num_margin}{The read numbe moves to the left.}

\item{show_transcript_name}{Whether the transcript name is displayed in the plot.}
}
\value{
return a ggplot2 object
}
\description{
Drawing the transcript exon and intron structure of multiple genes at a large scale.
}
\examples{
multi_gene(annotation,MUT1_read,readmap,chr='6',strand="both",range=c(36500000,36750000),limit=30,show_transcript_name=TRUE)
}
