% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_comp.R
\name{samples_comp}
\alias{samples_comp}
\title{samples_comp}
\usage{
samples_comp(
  annotation,
  readlist,
  readmap,
  interest_transcript_ids,
  novel_transcript_ids = NULL,
  limit = 30,
  read_color = "#ADD8E6",
  novel_isoform_color = "#CC0000",
  known_isoform_color = "#00DD00",
  text_color = "#000000",
  text_size = 3,
  text_alpha = 1,
  text_fontface = "bold",
  title_color = "#000000",
  title_size = 10,
  title_face = "bold",
  x_text_size = 5,
  num_color = "#000000",
  num_size = 3,
  num_alpha = 1,
  num_fontface = "bold",
  num_margin = 100,
  show_transcript_name = FALSE
)
}
\arguments{
\item{annotation}{A gtf format data,such as Homo_sapiens.GRCh38.109.chr.gtf.}

\item{readlist}{A list object with the read data of multiple samples.}

\item{readmap}{A tsv format data,mapping the transcript_id to read name;The column name needs to contain the name and transcript id.}

\item{interest_transcript_ids}{A dataframe with the column name needs to contain interest_transcript_id to choose what you want to show transcript isoform.}

\item{novel_transcript_ids}{A dataframe with the column name needs to contain novel_transcript_id to mark the color of novel isoform.}

\item{limit}{To limit the number of read displays. If the read count exceeds the limit value, the limit+log2(read count-limit)  number of reads will be randomly selected.}

\item{read_color}{Displays the color of the read structure.}

\item{novel_isoform_color}{Mark the color of the novel transcript isoform.}

\item{known_isoform_color}{Mark the color of the known transcript isoform.}

\item{text_color}{The font color of transcript name.}

\item{text_size}{The font size of transcript name.}

\item{text_alpha}{The font transparency of transcript name.}

\item{text_fontface}{The  font shape of transcript name.}

\item{title_color}{The font color  of tittle name.}

\item{title_size}{The font size of tittle name.}

\item{title_face}{The  font shape of tittle name.}

\item{x_text_size}{The text size of x axis.}

\item{num_color}{The font color of read number.}

\item{num_size}{The font size of read number.}

\item{num_alpha}{The font transparency of read number.}

\item{num_fontface}{The  font shape of read number.}

\item{num_margin}{The read numbe moves to the left.}

\item{show_transcript_name}{Whether the transcript name is displayed in the plot.}
}
\value{
return a ggplot2 object
}
\description{
Comparison of transcript isoforms structural differences between multiple samples.
}
\examples{
readlist=list("MUT1"=MUT1_read,"MUT2"=MUT2_read,"WT"=WT_read)
interest_transcript_ids=data.frame(interest_transcript_id=c("ENST00000405375","ENST00000615513"))
samples_comp(annotation=annotation,readlist=readlist,readmap=readmap,limit=100,interest_transcript_ids=interest_transcript_ids)
}
