% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_transcript.R
\name{single_transcript}
\alias{single_transcript}
\title{single_transcript}
\usage{
single_transcript(
  annotation,
  read,
  readmap,
  interest_transcript_id,
  read_color = "#ADD8E6",
  isoform_color = "#00DD00",
  limit = 30,
  title_color = "#000000",
  title_size = 10,
  title_face = "bold",
  num_size = 10,
  x_text_size = 5,
  show_transcript_name = FALSE
)
}
\arguments{
\item{annotation}{A gtf format data,such as Homo_sapiens.GRCh38.109.chr.gtf.}

\item{read}{A bed format data after alignment with the genome,having the GRanges object and the metadata of name column.}

\item{readmap}{A tsv format data,mapping the transcript_id to read name;The column name needs to contain the name and transcript id.}

\item{interest_transcript_id}{A string data as the interest transcript_id.}

\item{read_color}{Displays the color of the read structure.}

\item{isoform_color}{Displays the color of the transcript isoform structure}

\item{limit}{To limit the number of read displays. If the read count exceeds the limit value,the limit number of reads will be randomly selected.}

\item{title_color}{The font color  of tittle name.}

\item{title_size}{The font size of tittle name.}

\item{title_face}{The  font shape of tittle name.}

\item{num_size}{The font size of read number.}

\item{x_text_size}{The text size of x axis.}

\item{show_transcript_name}{Whether the transcript name is displayed in the plot.}
}
\value{
return a ggplot2 object
}
\description{
Drawing the exon and intron structure of interest transcript isoform.
}
\examples{
single_transcript(annotation,MUT1_read,readmap,"ENST00000405375",show_transcript_name=TRUE)
}
